The end

+++ BlackJack/BlackJackCalculator.cpp
@@ -16,7 +16,7 @@ namespace nagisakuya {
 					for (int k = 10; k-- > 0; )
 					{
 						temp = WhattoDo(((Deck(1) - i) - j) - k, PlayerHand({ i,j }), DealerHand({ k }));
-						sum += temp.second * (i == j ? 1 : 2) * (double)(deck.count(i) * deck.count(j) * deck.count(k));
+						sum += temp.second * (i == j ? 1 : 2) * ((double)deck.count(i) * (double)deck.count(j) * (double)deck.count(k));
 					}
 				}
 			}
